requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic-settings
openai
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
python-dotenv
httpx
redis[asyncio]
apscheduler
